+++ CHANGES.rst
@@ -61,7 +61,7 @@ Minor changes
   is now always visible when scrolling the table. :pr:`1102` by :user:`Jérôme
   Dockès <jeromedockes>`.
 
-* Added a `DropNullColumn` transformer that drops columns that contain only null
+* Added a `DropColumnIfNull` transformer that drops columns that contain only null
   values. :pr:`1115` by :user: `Riccardo Cappuzzo <riccardocappuzzo>`
 
 Bug fixes